requirements.txt
streamlit==1.38.0
opencv-python-headless==4.9.0.80
Pillow==10.4.0
PyYAML==6.0.2
numpy==1.26.4
